Organizing the code a bit better and filling in some gaps. - edge-config.js for Edge-specific gap fills - baas-config.js for BaaS-specific config code. - one gulpfile.js for all the tasks - streetcarts-seed.js for seeding code - moved config JSON into a .json file (instead of the .js files).
