--- gulp/tasks/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "streetcarts",
-  "version": "0.0.1",
-  "description": "Full application integrating Apigee Edge proxies with an Apigee API BaaS data store.",
-  "main": "index.js",
-  "directories": {
-    "test": "tests"
-  },
-  "scripts": {
-    "test": "mocha "
-  },
-  "dependencies": {
-    "async": "*",
-    "gulp-util": "^3.0.7",
-    "q": "*"
-  },
-  "author": "straut@google.com",
-  "license": "ISC"
-}
